+++ tests/testthat/test-edge.R
@@ -1,6 +1,6 @@
 context("test-edge.R")
 
 test_that("edge cases", {
-  expect_error(cellnumbers(ghrsst, matrix(1)))
-  expect_error(cellnumbers(ghrsst, 1), "no method for 'query' of type numeric")
+  expect_error(expect_warning(cellnumbers(ghrsst, matrix(1))))
+  expect_error(expect_warning(cellnumbers(ghrsst, 1)), "no method for 'query' of type numeric")
 })